PEP8

+++ test_app.py
@@ -42,33 +42,28 @@ class CapstoneTestCase(unittest.TestCase):
     Endpoint Test
     '''
     def test_get_actors(self):
-        res = self.client().get('/actors', headers={'\
-            Authorization': f'Bearer {executive_producer}'})
+        res = self.client().get('/actors', headers={'Authorization': f'Bearer {executive_producer}'})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 200)
         self.assertEqual(data['success'], True)
 
     def test_get_movies(self):
-        res = self.client().get('/movies', headers={'\
-            Authorization': f'Bearer {executive_producer}'})
+        res = self.client().get('/movies', headers={'Authorization': f'Bearer {executive_producer}'})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 200)
         self.assertEqual(data['success'], True)
 
     def test_post_actor(self):
-        res = self.client().post('/actors', headers={'\
-            Authorization': f'Bearer {executive_producer}'}, json={"\
-                name": "test", "age": 1, "gender": "male"})
+        res = self.client().post('/actors', headers={'Authorization': f'Bearer {executive_producer}'}, json={"name": "test", "age": 1, "gender": "male"})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 200)
         self.assertEqual(data['success'], True)
 
     def test_error_post_actor(self):
-        res = self.client().post('/actors', headers={'\
-            Authorization': f'Bearer {executive_producer}'}, json={})
+        res = self.client().post('/actors', headers={'Authorization': f'Bearer {executive_producer}'}, json={})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 422)
@@ -76,17 +71,14 @@ class CapstoneTestCase(unittest.TestCase):
         self.assertEqual(data['error'], 422)
 
     def test_post_movie(self):
-        res = self.client().post('/movies', headers={'\
-            Authorization': f'Bearer {executive_producer}'}, json={"\
-                title": "test", "release_date": "2021-11-11"})
+        res = self.client().post('/movies', headers={'Authorization': f'Bearer {executive_producer}'}, json={"title": "test", "release_date": "2021-11-11"})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 200)
         self.assertEqual(data['success'], True)
 
     def test_error_post_movie(self):
-        res = self.client().post('/movies', headers={'\
-            Authorization': f'Bearer {executive_producer}'}, json={})
+        res = self.client().post('/movies', headers={'Authorization': f'Bearer {executive_producer}'}, json={})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 422)
@@ -96,9 +88,7 @@ class CapstoneTestCase(unittest.TestCase):
     def test_patch_actor(self):
         self.actor = Actor(name="test", age=1, gender="male")
         self.actor.insert()
-        res = self.client().patch(f'/actors/{self.actor.id}', headers={'\
-            Authorization': f'Bearer {executive_producer}'}, json={"\
-                name": "test1"})
+        res = self.client().patch(f'/actors/{self.actor.id}', headers={'Authorization': f'Bearer {executive_producer}'}, json={"name": "test1"})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 200)
@@ -106,9 +96,7 @@ class CapstoneTestCase(unittest.TestCase):
         self.actor.delete()
 
     def test_error_patch_actor(self):
-        res = self.client().patch('/actors/999', headers={'\
-            Authorization': f'Bearer {executive_producer}'}, json={"\
-                name": "test1"})
+        res = self.client().patch('/actors/999', headers={'Authorization': f'Bearer {executive_producer}'}, json={"name": "test1"})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 404)
@@ -118,9 +106,7 @@ class CapstoneTestCase(unittest.TestCase):
     def test_patch_movie(self):
         self.movie = Movie(title="test", release_date="2021-11-11")
         self.movie.insert()
-        res = self.client().patch(f'/movies/{self.movie.id}', headers={'\
-            Authorization': f'Bearer {executive_producer}'}, json={"\
-                title": "test1"})
+        res = self.client().patch(f'/movies/{self.movie.id}', headers={'Authorization': f'Bearer {executive_producer}'}, json={"title": "test1"})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 200)
@@ -128,9 +114,7 @@ class CapstoneTestCase(unittest.TestCase):
         self.movie.delete()
 
     def test_error_patch_movie(self):
-        res = self.client().patch('/movies/999', headers={'\
-            Authorization': f'Bearer {executive_producer}'}, json={"\
-                title": "test1"})
+        res = self.client().patch('/movies/999', headers={'Authorization': f'Bearer {executive_producer}'}, json={"title": "test1"})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 404)
@@ -140,16 +124,14 @@ class CapstoneTestCase(unittest.TestCase):
     def test_delete_actor(self):
         self.actor = Actor(name="test", age=1, gender="male")
         self.actor.insert()
-        res = self.client().delete(f'/actors/{self.actor.id}', headers={'\
-            Authorization': f'Bearer {executive_producer}'})
+        res = self.client().delete(f'/actors/{self.actor.id}', headers={'Authorization': f'Bearer {executive_producer}'})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 200)
         self.assertEqual(data['success'], True)
 
     def test_error_delete_actor(self):
-        res = self.client().delete('/actors/999', headers={'\
-            Authorization': f'Bearer {executive_producer}'})
+        res = self.client().delete('/actors/999', headers={'Authorization': f'Bearer {executive_producer}'})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 404)
@@ -159,8 +141,7 @@ class CapstoneTestCase(unittest.TestCase):
     def test_delete_movie(self):
         self.movie = Movie(title="test", release_date="2021-11-11")
         self.movie.insert()
-        res = self.client().delete(f'/movies/{self.movie.id}', headers={'\
-            Authorization': f'Bearer {executive_producer}'})
+        res = self.client().delete(f'/movies/{self.movie.id}', headers={'Authorization': f'Bearer {executive_producer}'})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 200)
@@ -168,8 +149,7 @@ class CapstoneTestCase(unittest.TestCase):
         self.movie.delete()
 
     def test_error_delete_movie(self):
-        res = self.client().patch('/movies/999', headers={'\
-            Authorization': f'Bearer {executive_producer}'})
+        res = self.client().patch('/movies/999', headers={'Authorization': f'Bearer {executive_producer}'})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 404)
@@ -181,8 +161,7 @@ class CapstoneTestCase(unittest.TestCase):
     '''
 # view actors
     def test_casting_assistant(self):
-        res = self.client().get('/actors', headers={'\
-            Authorization': f'Bearer {casting_assistant}'})
+        res = self.client().get('/actors', headers={'Authorization': f'Bearer {casting_assistant}'})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 200)
@@ -190,9 +169,7 @@ class CapstoneTestCase(unittest.TestCase):
 
 # add actor
     def test_error_casting_assistant(self):
-        res = self.client().post('/actors', headers={'\
-            Authorization': f'Bearer {casting_assistant}'}, json={"\
-                name": "test", "age": 1, "gender": "male"})
+        res = self.client().post('/actors', headers={'Authorization': f'Bearer {casting_assistant}'}, json={"name": "test", "age": 1, "gender": "male"})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 401)
@@ -200,9 +177,7 @@ class CapstoneTestCase(unittest.TestCase):
 
 # add actor
     def test_casting_director(self):
-        res = self.client().post('/actors', headers={'\
-            Authorization': f'Bearer {casting_director}'}, json={"\
-                name": "test", "age": 1, "gender": "male"})
+        res = self.client().post('/actors', headers={'Authorization': f'Bearer {casting_director}'}, json={"name": "test", "age": 1, "gender": "male"})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 200)
@@ -210,9 +185,7 @@ class CapstoneTestCase(unittest.TestCase):
 
 # add movie
     def test_error_casting_director(self):
-        res = self.client().post('/movies', headers={'\
-            Authorization': f'Bearer {casting_director}'}, json={"\
-                title": "test", "release_date": "2021-11-11"})
+        res = self.client().post('/movies', headers={'Authorization': f'Bearer {casting_director}'}, json={"title": "test", "release_date": "2021-11-11"})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 401)
@@ -220,9 +193,7 @@ class CapstoneTestCase(unittest.TestCase):
 
 # add movie
     def test_executive_producer(self):
-        res = self.client().post('/movies', headers={'\
-            Authorization': f'Bearer {executive_producer}'}, json={"\
-                title": "test", "release_date": "2021-11-11"})
+        res = self.client().post('/movies', headers={'Authorization': f'Bearer {executive_producer}'}, json={"title": "test", "release_date": "2021-11-11"})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 200)
@@ -232,9 +203,7 @@ class CapstoneTestCase(unittest.TestCase):
     def test_executive_producer_2(self):
         self.movie = Movie(title="test", release_date="2021-11-11")
         self.movie.insert()
-        res = self.client().delete(f'\
-            /movies/{self.movie.id}', headers={'\
-                Authorization': f'Bearer {executive_producer}'})
+        res = self.client().delete(f'/movies/{self.movie.id}', headers={'Authorization': f'Bearer {executive_producer}'})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 200)
